StylePlugin tests 100%
